% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{betahatR}
\alias{betahatR}
\title{Estimate the coefficients of the NAR(1).}
\usage{
betahatR(Y, X)
}
\arguments{
\item{Y}{the samples generated by 'obR' function (numeric)}

\item{X}{the samples generated by 'obR' function (numeric)}
}
\value{
the estimated coefficients vector and a inverse matrix (numeric)
}
\description{
Estimate the coefficients of the NAR(1).
}
\examples{
\dontrun{
T<- 1000
K<- 3
N<- 10
beta<- c(0.2,-0.4)
A<- netA(K,N)
res<- obR(A,beta,T)
hatbeta<- betahatR(res$Y,res$X)$betahat}
}
