% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GibbsC}
\alias{GibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
GibbsC(N)
}
\arguments{
\item{N}{the number of samples}
}
\value{
a random sample (x,y)
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
GibbsC(100)
}
}
