% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{netA}
\alias{netA}
\title{Generate NAR(1) model samples using R.}
\usage{
netA(K, N)
}
\arguments{
\item{K}{the number of blocks (integer)}

\item{N}{the network size (integer)}
}
\value{
the network structure matrix A
}
\description{
Generate NAR(1) model samples using R.
}
\examples{
\dontrun{
K<- 3
N<- 5
netA(3,5)}
}
