% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{obR}
\alias{obR}
\title{Generate the sample Y_t of NAR(1) model.}
\usage{
obR(A, beta, T)
}
\arguments{
\item{A}{network structure matrix A}

\item{beta}{the 2 dimensional vector of coefficient (numeric)}

\item{T}{the sample size (integer)}
}
\value{
the sample list of (X,Y) (numeric)
}
\description{
Generate the sample Y_t of NAR(1) model.
}
\examples{
\dontrun{
T<- 1000
K<- 5
N<- 100
beta<- c(0.2,-0.4)
A<- StatComp21024::netA(K,N)
obR(A,beta,T)}
}
