% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{RMSE}
\alias{RMSE}
\title{Compute the RMSE (root mean square error) and CP (coverage probability).}
\usage{
RMSE(rep, K, N, beta, T)
}
\arguments{
\item{rep}{the number of replications (integer)}

\item{K}{the number of blocks (integer)}

\item{N}{the network size (integer)}

\item{beta}{the 2 dimensional vector of coefficient (numeric)}

\item{T}{the sample size (integer)}
}
\value{
the estimated coefficients vector (numeric)
}
\description{
Compute the RMSE (root mean square error) and CP (coverage probability).
}
\examples{
\dontrun{
rep<- 1000
T<- 100
K<- 3
N<- 10
beta<- c(0.2,-0.4)
mseCPR(rep,K,N,beta,T)}
}
