\name{StatComp21024-package}
\alias{StatComp21024-package}
\alias{StatComp21024}
\docType{package}
\title{
  Use Stochastic Block Model (SBM) to construct network adjacency matrix A and estimate coefficients of NAR(1) model.
}
\description{
  Use Stochastic Block Model (SBM) to construct network adjacency matrix A, estimate coefficients of NAR(1) model and simulates two measures to gauge estimators' performances.
  (\code{netA},\code{obR},\code{betahatR} and \code{RMSE})
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Lu Cheng

Maintainer: Lu Cheng <clulu991014@mail.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
